Rename components to reflect their functionality more accurately.
