+++ src/Components/BehavioralExecutionTree/BehavioralExecutionNode/BehavioralExecutionNode.js
@@ -1,14 +1,14 @@
 import React, {useContext, useEffect, useState} from "react";
 
 import PropTypes from "prop-types";
-import {CaretDownFill, CaretRightFill, Stack} from "react-bootstrap-icons";
+import {CaretDownFill, CaretRightFill, SignpostFill, Stack} from "react-bootstrap-icons";
 
 import BreakpointsContext from "../../../Providers/BreakpointsContext";
-import DesignTreeInstanceContext from "../DesignTreeInstanceContext";
+import ExecutionTreeInstanceContext from "../BehavioralExecutionTreeInstanceContext";
 
-import "./DesignNode.scss";
+import "./BehavioralExecutionNode.scss";
 
-DesignNode.propTypes = {
+BehavioralExecutionNode.propTypes = {
     node: PropTypes.object,
 };
 
@@ -17,60 +17,42 @@ DesignNode.propTypes = {
  * @param {Object} node
  * @return {JSX.Element}
  */
-export function DesignNode ({node}) {
-    const {selectedNode, selectNode,
-        toggleCollapse, setCollapsed} = useContext(DesignTreeInstanceContext);
-    const [selectedStyle, setSelectedStyle] = useState();
-    const [hasViolation, setHasViolation] = useState();
-    const [debugText, setDebugText] = useState();
+export function BehavioralExecutionNode ({node}) {
     const {breakPoints} = useContext(BreakpointsContext);
+    const {selectedNode, selectNode, toggleCollapse} = useContext(ExecutionTreeInstanceContext);
+    const [selectedStyle, setSelectedStyle] = useState();
+    const [debugText, setDebugText] = useState();
+    const [hasViolation, setHasViolation] = useState();
 
     // Set style if node is selected.
     useEffect(() => {
-        if (node) {
-            for (let i = 0; i < node.execution.length; i++) {
-                const exec = node.execution[i];
-                if ("violation" in exec) {
-                    setHasViolation(true);
-                    setDebugText("Root Cause");
-                }
-            }
-        }
         if (node && selectedNode) {
-            if (node.exception) {
-                setSelectedStyle({background: "#3f191b", color: "white"});
-            } else if (selectedNode === node) {
-                setSelectedStyle({background: "#003d35", color: "white"});
+            // console.log(node.designAbstraction?.functionalAbs?.type);
+            if (selectedNode === node) {
+                setSelectedStyle({background: "#3b3b3b", color: "white"});
             } else {
                 setSelectedStyle({});
             }
         }
-    }, [selectedNode, node]);
+        if (node) {
+            setHasViolation(false);
+            setDebugText(undefined);
 
-
-    const getBreakPoint = () => {
-        if (!breakPoints) return;
-        let isEnabled;
-        let hasBreakPoint;
-        for (let k = 0; k < node.execution.length; k++) {
-            const entry = node.execution[k];
-            for (let i = 0; i < breakPoints.length; i++) {
-                const point = breakPoints[i];
-                if (point.abstraction_meta === entry.abstractionId) {
-                    hasBreakPoint = true;
-                    if (point.enabled) {
-                        isEnabled = true;
-                    }
-                }
+            if (node.invalid) {
+                setHasViolation(true);
+                setDebugText("violation");
+            }
+            if (node.rootCause) {
+                setHasViolation(true);
+                setDebugText("root cause");
+            }
+            if (node.exception) {
+                setHasViolation(true);
+                setSelectedStyle({background: "#3f191b", color: "white"});
+                setDebugText("failure");
             }
         }
-        if (hasBreakPoint) {
-            const className = (isEnabled == true)?
-                "enabledBreakPoint":
-                "disabledBreakPoint";
-            return <div className={className}></div>;
-        }
-    };
+    }, [selectedNode, node]);
 
     /**
      * Callback when a node is toggled.
@@ -79,7 +61,6 @@ export function DesignNode ({node}) {
      */
     const clickToggle = (e, node) => {
         e.preventDefault();
-        e.stopPropagation();
         if (node.collapsible) {
             toggleCollapse(node);
         }
@@ -93,7 +74,6 @@ export function DesignNode ({node}) {
     const clickSelectNode = (e, node) => {
         e.preventDefault();
         selectNode(node);
-        setCollapsed(node, false);
     };
 
     /**
@@ -113,7 +93,6 @@ export function DesignNode ({node}) {
         }
     };
 
-
     /**
      * Creates space for each node level.
      * @param {Object} node
@@ -121,7 +100,7 @@ export function DesignNode ({node}) {
      */
     const getSpacers = (node) => {
         const spacers = [];
-        for (let i = 0; i < node.level; i++) {
+        for (let i = 0; i < node.level - 1; i++) {
             spacers.push(
                 <div className="spacer" key={i}>
                     <div className="vertical-line"></div>
@@ -136,10 +115,28 @@ export function DesignNode ({node}) {
      * @return {JSX}
      */
     const getNodeIconType = () => {
-        if (node.isBehavior) {
+        if (node.abstractionType === "function_call") {
             return <Stack
                 title="Function Call"
                 style={{color: "orange"}}/>;
+        } else if (node.abstractionType === "conditional_branch") {
+            return <SignpostFill
+                title="Conditional Branch"
+                style={{color: "#3794ff"}}/>;
+        }
+    };
+
+
+    const getBreakPoint = () => {
+        if (!breakPoints) return;
+        for (let i = 0; i < breakPoints.length; i++) {
+            const point = breakPoints[i];
+            if (point.abstraction_meta === node.abstractionId) {
+                const className = (point.enabled == true)?
+                    "enabledBreakPoint":
+                    "disabledBreakPoint";
+                return <div className={className}></div>;
+            }
         }
     };
 
@@ -168,16 +165,9 @@ export function DesignNode ({node}) {
                     {getCollapsed(node)}
                 </div>
 
-
-                {
-                    node.behavior ?
-                        <div className="text-container flex-grow-1">
-                            <span>{node.intent}</span>
-                        </div>:
-                        <div className="text-container flex-grow-1">
-                            <span>{node.entry.id}</span>
-                        </div>
-                }
+                <div className="text-container flex-grow-1">
+                    <span>{node.intent}</span>
+                </div>
 
                 {hasViolation ?
                     <div className="analysis-status-container">
@@ -185,7 +175,6 @@ export function DesignNode ({node}) {
                     </div>:
                     <></>
                 }
-
             </div>
         </div>
     );
