+++ src/Components/BehavioralExecutionTree/BehavioralExecutionNode/BehavioralExecutionNode.scss
