+++ src/Components/BehavioralExecutionTree/BehavioralExecutionTreeInstanceContext.js
@@ -1,0 +1,6 @@
+import {createContext} from "react";
+
+// This context stores the execution tree instance information.
+const BehavioralExecutionTreeInstanceContext = createContext();
+
+export default BehavioralExecutionTreeInstanceContext;
