--- src/Components/ExecutionTree/ExecutionTreeInstanceContext.js
@@ -1,6 +1,0 @@
-import {createContext} from "react";
-
-// This context stores the execution tree instance information.
-const ExecutionTreeInstanceContext = createContext();
-
-export default ExecutionTreeInstanceContext;
